requirements.txt
streamlit
pandas
plotly
xgboost
scikit-learn
requests
ollama
numpy
matplotlib
seaborn
requests
tqdm
jupyter
IPython
pathlib

bcrypt
python-dotenv
uvicorn
starlette
itsdangerous
authlib